requirements.txt
PyGithub==2.4.0
pydantic==2.8.2
pandas>=2.0.0
scikit-learn>=1.3.0
streamlit>=1.30.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0
GitPython>=3.1.0
